Added assertion to match sub claim in proxy header and jwt assertion. Updated 2 other unit tests.
