+++ server/src/main/java/org/cloudfoundry/identity/uaa/provider/token/ClientAssertionHeaderAuthenticator.java
@@ -1,7 +1,10 @@
 package org.cloudfoundry.identity.uaa.provider.token;
 
+import java.util.Map;
+
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
+import org.cloudfoundry.identity.uaa.oauth.token.ClaimConstants;
 import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.security.jwt.Jwt;
 import org.springframework.security.jwt.JwtHelper;
@@ -17,15 +20,15 @@ public class ClientAssertionHeaderAuthenticator {
      * @throws BadCredentialsException must throw this if authentication fails
      */
     public String authenticate(final String clientAssertionHeader, final String proxyPublicKey) throws BadCredentialsException {
-        Jwt jwt = null;
+        Jwt clientAssertion = null;
         try {
             if (StringUtils.hasText(clientAssertionHeader)) {
-                jwt = JwtHelper.decode(clientAssertionHeader);
-                jwt.verifySignature(new RsaVerifier(proxyPublicKey));
-                return jwt.getClaims();
+                clientAssertion = JwtHelper.decode(clientAssertionHeader);
+                clientAssertion.verifySignature(new RsaVerifier(proxyPublicKey));
+                return clientAssertion.getClaims();
             }
         } catch (RuntimeException e) {
-            logger.debug("Validation failed for client assertion header. Header:{" + jwt + "}; error: " + e);
+            logger.debug("Validation failed for client assertion header. Header:{" + clientAssertion + "}; error: " + e);
         }
 
         // Do not include error detail in this exception.
